# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Dimbaa Backend Api Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://dimbaa-be.test'
paths:
...
  /api/organizers/matchrecords:
    get:
      summary: 'Display a listing of the resource.'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: 'Store a newly created resource in storage.'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tournament:
                  type: integer
                  description: ''
                  example: 9
                date:
                  type: string
                  description: 'Must be a valid date. Must be a date after <code>yesterday</code>.'
                  example: '2035-02-22'
                home_team:
                  type: integer
                  description: ''
                  example: 9
                away_team:
                  type: integer
                  description: ''
                  example: 9
                stadium:
                  type: integer
                  description: ''
                  example: 8
                city:
                  type: string
                  description: "'. Must be between 3 and 60 characters."
                  example: akv
                round:
                  type: integer
                  description: 'Must be between 1 and 100.'
                  example: 1
              required:
                - tournament
                - date
                - home_team
                - away_team
                - stadium
                - city
                - round
      security: []
  '/api/organizers/matchrecords/{id}':
    get:
      summary: 'Display the specified resource.'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    put:
      summary: 'Update the specified resource in storage.'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      security: []
    delete:
      summary: 'Remove the specified resource from storage.'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the matchrecord.'
        example: molestias
        required: true
        schema:
          type: string
...